requirements.txt
Django
django-rest-framework
gunicorn
drf-spectacular
psycopg2
djangorestframework-simplejwt
satella